Exercise 2 : Validate The Email
We will create a form that ask the user for his email. Then, using a function we will check the validity of the user’s input.
1. Add an input of type “email” to your form.
2. Write your own javascript validation function to check if the entered value is a valid email address.
I.e contains some valid characters, and @ sign, more characters a . (dot) and some more characters.
3. On “submit”, the validation function should run and validate the email address.
(Do not use a built in method or regex).